:recycle: refactor : refector layout
separate header
