+++ src/components/common/header/_component/ShowSearchCart.tsx
@@ -1,0 +1,55 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import { useState } from 'react';
+import { GoSearch } from 'react-icons/go';
+import { PiShoppingCartSimple } from 'react-icons/pi';
+import SearchBar from './SearchBar';
+
+interface ShowSearchCartProps {
+  showSearch: boolean;
+  showCart: boolean;
+}
+
+const ShowSearchCart = ({ showSearch, showCart }: ShowSearchCartProps) => {
+  const router = useRouter();
+  const [isSearchVisible, setIsSearchVisible] = useState(false);
+
+  const handleSearchToggle = () => {
+    setIsSearchVisible((prev) => !prev);
+  };
+
+  // 검색, 카트 있다 = 기본 / 검색만 있음 / 검색, 카트 둘 다 없음 = false
+  if (showSearch && showCart) {
+    return (
+      <div className="flex p-1 gap-1">
+        {isSearchVisible && <SearchBar />}
+        <GoSearch className="w-7 h-7" onClick={handleSearchToggle} />
+        <PiShoppingCartSimple
+          onClick={() => router.push('/cart')}
+          className="cursor-pointer w-7 h-7"
+        />
+      </div>
+    );
+  } else if (showSearch && !showCart) {
+    return (
+      <div className="flex p-1">
+        {isSearchVisible && <SearchBar />}
+        <GoSearch className="w-7 h-7" onClick={handleSearchToggle} />
+      </div>
+    );
+  } else if (!showSearch && showCart) {
+    return (
+      <div className="flex p-1">
+        <PiShoppingCartSimple
+          onClick={() => router.push('/cart')}
+          className="cursor-pointer w-7 h-7"
+        />
+      </div>
+    );
+  } else {
+    return <div className="invisible w-7 h-7" />;
+  }
+};
+
+export default ShowSearchCart;
