--- src/components/search/Search.tsx
@@ -1,5 +1,0 @@
-const Search = () => {
-  return <div>Search</div>;
-};
-
-export default Search;
